moved logic for finding web elements into a separate class
