+++ HTTPDriver/HTTPDriver.csproj
@@ -62,6 +62,7 @@
     <Compile Include="Navigation.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="WebElement.cs" />
+    <Compile Include="WebElementFinder.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
